style: use jss to write dynamic theme
