+++ packages/theme/src/themeRelate/index.ts
@@ -1,49 +1,20 @@
-import type { Scheme } from '@material/material-color-utilities'
-import {
-  applyTheme,
-  argbFromHex,
-  hexFromArgb,
-  themeFromSourceColor,
-} from '@material/material-color-utilities'
-import state from '../state'
+import jss from 'jss'
+import preset from 'jss-preset-default'
 import color from '../color'
+import state from '../state'
 import type FnTheme from './theme'
-import type { Palette } from './theme'
+import scheme from './scheme'
 
-const dynamicTheme = themeFromSourceColor(argbFromHex('#3894ff'))
-const systemDark = window.matchMedia('(prefers-color-scheme: dark)').matches
-
-applyTheme(dynamicTheme, { target: document.documentElement, dark: systemDark })
-
-// const toKebabize = (str: string) =>
-//   str.replace(
-//     /[A-Z]+(?![a-z])|[A-Z]/g,
-//     ($, ofs) => (ofs ? '-' : '') + $.toLowerCase()
-//   )
-
-// export const createTheme = (theme: ThemeConfig) => {
-//   const { palette } = theme
-//   const curPalette = Object.keys(palette).reduce((pre, cur) => {
-//     const key = `--md-sys-color-${toKebabize(cur)}`
-//     pre[key] = palette[cur as keyof PaletteConfig]
-//     return pre
-//   }, {} as any)
-//   return { ...curPalette }
-// }
-
-export const parseShceme = (scheme: Scheme): Palette => {
-  const newScheme: Record<keyof Scheme, string> = {} as any
-  for (const [key, value] of Object.entries(scheme.toJSON())) {
-    newScheme[key as keyof Scheme] = hexFromArgb(value)
-  }
-
-  return newScheme
+jss.setup(preset())
+const styles = {
+  '@global': {
+    ':root': scheme.styles,
+  },
 }
+jss.createStyleSheet(styles).attach()
 
 const theme: FnTheme = {
-  palette: systemDark
-    ? parseShceme(dynamicTheme.schemes.dark)
-    : parseShceme(dynamicTheme.schemes.light),
+  palette: scheme.palette,
   color,
   state,
 }
